moved starboard to new cog as it got bigger, delete entry command, delete all starboard entries command, should work ish need stress test
